moving cache code to its own package
