+++ crewai/agents/__init__.py
@@ -1,4 +1,4 @@
-from .cache_handler import CacheHandler
+from .cache.cache_handler import CacheHandler
 from .executor import CrewAgentExecutor
 from .output_parser import CrewAgentOutputParser
 from .tools_handler import ToolsHandler
